Remove SelectedCompoundPairs and replace tests with tests for SelectedInchikeyPairs
