+++ ms2deepscore/train_new_model/__init__.py
@@ -1,10 +1,10 @@
 from .data_generators import DataGeneratorPytorch
-from .inchikey_pair_selection import (SelectedCompoundPairs,
+from .inchikey_pair_selection import (SelectedInchikeyPairs,
                                       select_compound_pairs_wrapper)
 
 
 __all__ = [
     "DataGeneratorPytorch",
     "select_compound_pairs_wrapper",
-    "SelectedCompoundPairs",
+    "SelectedInchikeyPairs",
 ]
